Made the read_data function (and its supproting translation function) into a DataReader class.  It is instantiated with a filename that contains the list of csv files. To get the dino filters, object we can generate and instance of the DataReader class and call the read_data function.
